Renamed matchmaking module to lobby
