+++ src/lobby/Authenticator.js
