+++ src/lobby/Matchmaker.js
@@ -1,4 +1,4 @@
-import '../styles/matchmaking.css';
+import '../styles/lobby.css';
 import React from 'react';
 import { LobbyClient } from 'boardgame.io/client';
 import Lobby from './Lobby'; // manages game creation and joining
